Add files via upload
This is a simple Basic Banking Webpage project that I built using HTML, CSS, and JavaScript. It is a responsive website that can be used to simulate banking transactions and operations, such as creating an account, transferring money, and viewing account details. The website features a clean and user-friendly interface that allows users to easily navigate through its pages and perform the necessary actions. This project is perfect for beginners who are looking to practice their web development skills or anyone who wants to learn more about the basics of building a functional banking website.